moved passport stuff to another file
