+++ src/server/server.js
@@ -4,7 +4,6 @@ const express = require("express");
 const app = express();
 
 // DB setup
-const userSchema = require("../models/users");
 const mongoose = require("mongoose");
 
 // Our Middleware 
@@ -16,7 +15,7 @@ const session = require("express-session");
 const sessionStore = require('connect-mongodb-session')(session);
 const cors = require('cors')
 const passport = require('passport')
-const LocalStrategy =  require('passport-local').Strategy;
+const localStrategy = require('./passport');
 
 //Database config
 const db = `mongodb+srv://${env.DB_USER}:${env.DB_PASSWORD}@${env.DB_CLUSTER}/${env.LOGIN_DB}?retryWrites=true&w=majority`;
@@ -55,32 +54,7 @@ app.use(session({ //setup session middleware
 })); 
 app.use(passport.initialize()); // initialize passport + sessions
 app.use(passport.session()); // initialize passport + sessions
-
-passport.use(new LocalStrategy(
-  function(username, password, done) {
-    userSchema.findOne({ username: username }, 
-      function (err, user) {
-        if (err) { return done(err)}; 
-        if (!user) { 
-          alogator('yellow', `${env.SERVER_NAME} User '${username}' does not exist in the database. Login denied.`)
-          return done(null, false); }
-        if (user.password !== password) {
-          alogator('yellow', `${env.SERVER_NAME} User '${username}' has entered an incorrect password. Login denied.`)
-          return done(null, false)
-        }
-        return done(null, user);
-      }
-    );
-  }
-));
-
-passport.serializeUser(function(user, done) {
-  done(null, user);
-});
-
-passport.deserializeUser(function(user, done) {
-  done(null, user);
-});
+passport.use(localStrategy);
 
 // *** Routes ***//
 app.post('/auth',
@@ -99,7 +73,7 @@ app.get("/authtest", isAuth(), (req, res) => {
 
 app.get("/admin", isAuth(), (req, res) => {
   if (req.user.accessLevel === 'owner') {
-    aLOGator("green" , "Administrator authorized");
+    alogator("green" , "Administrator authorized");
     res.status(200).send("Administrator authorized")
   } else {
     res.redirect('/failure')
